back_end/requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
logging